requirements..txt
discord.py
jaconv
pykakasi
flask